added more tests for SelTestDesc
