+++ pepper-framework/src/test/resources/selfTest/comparisonTests/whenComparingTwoCorpusPathesWithEqualFiles/expected/corpus1/corpus2/expected.txt
@@ -1,0 +1,2 @@
+
+Any content
