+++ pepper-framework/src/test/resources/selfTest/comparisonTests/whenComparingTwoCorpusPathesWithEqualFiles/expected/corpus1/corpus2/expected.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<a a="1" b="2">
+  <b>
+    <c/>
+  </b>
+</a>
